move keystore implementations into `keystore` package
This commit moves all external KMS keystore
implementations (Vault, AWS, GCP, ...) into
a dedicated `keystore` package.

This separates the stateless and stateful KES
implementation and makes it more clear where
to add things.
