+++ internal/keystore/azure/client.go
