+++ internal/keystore/azure/error.go
