+++ internal/keystore/generic/client.go
