+++ internal/keystore/generic/spec-v1.md
