+++ internal/keystore/vault/client.go
