+++ internal/keystore/vault/iterator.go
