+++ internal/keystore/vault/vault.go
