+++ keserv/config.go
@@ -13,15 +13,15 @@ import (
 	"time"
 
 	"github.com/minio/kes"
-	"github.com/minio/kes/internal/aws"
-	"github.com/minio/kes/internal/azure"
-	"github.com/minio/kes/internal/fortanix"
-	"github.com/minio/kes/internal/fs"
-	"github.com/minio/kes/internal/gcp"
-	"github.com/minio/kes/internal/gemalto"
-	"github.com/minio/kes/internal/generic"
-	"github.com/minio/kes/internal/mem"
-	"github.com/minio/kes/internal/vault"
+	"github.com/minio/kes/internal/keystore/aws"
+	"github.com/minio/kes/internal/keystore/azure"
+	"github.com/minio/kes/internal/keystore/fortanix"
+	"github.com/minio/kes/internal/keystore/fs"
+	"github.com/minio/kes/internal/keystore/gcp"
+	"github.com/minio/kes/internal/keystore/gemalto"
+	"github.com/minio/kes/internal/keystore/generic"
+	"github.com/minio/kes/internal/keystore/mem"
+	"github.com/minio/kes/internal/keystore/vault"
 	"github.com/minio/kes/kms"
 	"gopkg.in/yaml.v3"
 )
